+++ pkg/gio/v2/gcontenttype.go
@@ -495,7 +495,7 @@ func ContentTypeSetMIMEDirs(dirs []string) {
 	var _arg1 **C.gchar // out
 
 	{
-		_arg1 = (**C.gchar)(C.malloc(C.ulong(len(dirs)+1) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg1 = (**C.gchar)(C.malloc(C.size_t(len(dirs)+1) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg1))
 		{
 			out := unsafe.Slice(_arg1, len(dirs)+1)
