+++ pkg/gio/v2/gsocket.go
@@ -1428,7 +1428,7 @@ func (socket *Socket) ReceiveMessages(ctx context.Context, messages []InputMessa
 		_arg4 = (*C.GCancellable)(unsafe.Pointer(cancellable.Native()))
 	}
 	_arg2 = (C.guint)(len(messages))
-	_arg1 = (*C.GInputMessage)(C.malloc(C.ulong(len(messages)) * C.ulong(C.sizeof_GInputMessage)))
+	_arg1 = (*C.GInputMessage)(C.malloc(C.size_t(len(messages)) * C.size_t(C.sizeof_GInputMessage)))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.GInputMessage)(_arg1), len(messages))
@@ -1633,7 +1633,7 @@ func (socket *Socket) SendMessage(ctx context.Context, address SocketAddresser, 
 		_arg1 = (*C.GSocketAddress)(unsafe.Pointer(address.Native()))
 	}
 	_arg3 = (C.gint)(len(vectors))
-	_arg2 = (*C.GOutputVector)(C.malloc(C.ulong(len(vectors)) * C.ulong(C.sizeof_GOutputVector)))
+	_arg2 = (*C.GOutputVector)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_GOutputVector)))
 	defer C.free(unsafe.Pointer(_arg2))
 	{
 		out := unsafe.Slice((*C.GOutputVector)(_arg2), len(vectors))
@@ -1643,7 +1643,7 @@ func (socket *Socket) SendMessage(ctx context.Context, address SocketAddresser, 
 	}
 	if messages != nil {
 		_arg5 = (C.gint)(len(messages))
-		_arg4 = (**C.GSocketControlMessage)(C.malloc(C.ulong(len(messages)) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg4 = (**C.GSocketControlMessage)(C.malloc(C.size_t(len(messages)) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg4))
 		{
 			out := unsafe.Slice((**C.GSocketControlMessage)(_arg4), len(messages))
@@ -1716,7 +1716,7 @@ func (socket *Socket) SendMessageWithTimeout(ctx context.Context, address Socket
 		_arg1 = (*C.GSocketAddress)(unsafe.Pointer(address.Native()))
 	}
 	_arg3 = (C.gint)(len(vectors))
-	_arg2 = (*C.GOutputVector)(C.malloc(C.ulong(len(vectors)) * C.ulong(C.sizeof_GOutputVector)))
+	_arg2 = (*C.GOutputVector)(C.malloc(C.size_t(len(vectors)) * C.size_t(C.sizeof_GOutputVector)))
 	defer C.free(unsafe.Pointer(_arg2))
 	{
 		out := unsafe.Slice((*C.GOutputVector)(_arg2), len(vectors))
@@ -1726,7 +1726,7 @@ func (socket *Socket) SendMessageWithTimeout(ctx context.Context, address Socket
 	}
 	if messages != nil {
 		_arg5 = (C.gint)(len(messages))
-		_arg4 = (**C.GSocketControlMessage)(C.malloc(C.ulong(len(messages)) * C.ulong(unsafe.Sizeof(uint(0)))))
+		_arg4 = (**C.GSocketControlMessage)(C.malloc(C.size_t(len(messages)) * C.size_t(unsafe.Sizeof(uint(0)))))
 		defer C.free(unsafe.Pointer(_arg4))
 		{
 			out := unsafe.Slice((**C.GSocketControlMessage)(_arg4), len(messages))
@@ -1818,7 +1818,7 @@ func (socket *Socket) SendMessages(ctx context.Context, messages []OutputMessage
 		_arg4 = (*C.GCancellable)(unsafe.Pointer(cancellable.Native()))
 	}
 	_arg2 = (C.guint)(len(messages))
-	_arg1 = (*C.GOutputMessage)(C.malloc(C.ulong(len(messages)) * C.ulong(C.sizeof_GOutputMessage)))
+	_arg1 = (*C.GOutputMessage)(C.malloc(C.size_t(len(messages)) * C.size_t(C.sizeof_GOutputMessage)))
 	defer C.free(unsafe.Pointer(_arg1))
 	{
 		out := unsafe.Slice((*C.GOutputMessage)(_arg1), len(messages))
